fix: 修复组长以外的用户可以修改用户组的问题

+++ issuer/routers/user_group.py
@@ -61,7 +61,9 @@ async def delete_user_group(user_group_model: "UserGroupReq",
     _user = check_cookie(cookie=current_user)
     if _user is None:
         return {"success": False, "reason": "Invalid token"}
-    # TODO: 检查是否为用户组的owner
+    user_group = db.find_user_group_by_code(user_group_model.group_code)
+    if user_group.group_owner != _user.user_name:
+        return {"success": False, "reason": "Permission denied"}
     res = db.delete_user_group_by_code(user_group_model.group_code)
     if res is False:
         return {"success": res}
@@ -86,7 +88,9 @@ async def change_user_group(user_group_model: "UserGroupReq",
     _user = check_cookie(cookie=current_user)
     if _user is None:
         return {"success": False, "reason": "Invalid token"}
-    # TODO: 检查是否是用户组的owner，如果不是只能退出用户组
+    user_group = db.find_user_group_by_code(user_group_model.group_code)
+    if user_group.group_owner != _user.user_name:
+        return {"success": False, "reason": "Permission denied"}
     _group_code = user_group_model.group_code
     _user_group = db.find_user_group_by_code(_group_code)
     _user_group.group_name = user_group_model.group_name
